fix print

+++ code_gen.h
@@ -31,7 +31,7 @@ int gen_builtin_fun(char *fun_id);
 int gen_inputs();
 int gen_inputi();
 int gen_inputf();
-int gen_print(size_t params_count);
+int gen_print(unsigned params_count);
 int gen_length();
 int gen_substr();
 int gen_ord();
